refactor: 充足date公用方法
