+++ src/packages/calendar/calendarviewmodeitem.taro.tsx
@@ -2,19 +2,18 @@ import React, { useState, useEffect, useRef } from 'react'
 import classNames from 'classnames'
 import { ScrollView, View } from '@tarojs/components'
 import { ComponentDefaults } from '@/utils/typings'
+import { getDateString, compareDate } from '@/utils/date'
+import requestAniFrame from '@/utils/raf'
+import { useConfig } from '@/packages/configprovider'
+import { usePropsValue } from '@/utils/use-props-value'
 import {
-  getDateString,
-  compareDate,
+  splitDate,
   getPreMonths,
   getMonths,
   getPreQuarters,
   getNextQuarters,
   getQuarters,
-} from '@/utils/date'
-import requestAniFrame from '@/utils/raf'
-import { useConfig } from '@/packages/configprovider'
-import { usePropsValue } from '@/utils/use-props-value'
-import { splitDate } from './utils'
+} from './utils'
 import {
   CalendarDay,
   CalendarValue,
