+++ src/packages/calendar/calendarviewmodeitem.tsx
@@ -1,19 +1,18 @@
 import React, { useState, useEffect, useRef } from 'react'
 import classNames from 'classnames'
 import { ComponentDefaults } from '@/utils/typings'
+import { getDateString, compareDate } from '@/utils/date'
+import requestAniFrame from '@/utils/raf'
+import { useConfig } from '@/packages/configprovider'
+import { usePropsValue } from '@/utils/use-props-value'
 import {
-  getDateString,
-  compareDate,
+  splitDate,
   getPreMonths,
   getMonths,
   getPreQuarters,
   getNextQuarters,
   getQuarters,
-} from '@/utils/date'
-import requestAniFrame from '@/utils/raf'
-import { useConfig } from '@/packages/configprovider'
-import { usePropsValue } from '@/utils/use-props-value'
-import { splitDate } from './utils'
+} from './utils'
 import {
   CalendarDay,
   CalendarValue,
