+++ src/packages/calendaritem/calendaritem.tsx
@@ -5,18 +5,20 @@ import { PopupProps } from '@/packages/popup/index'
 import { ComponentDefaults } from '@/utils/typings'
 import {
   getDateString,
-  getCurrMonthData,
-  getDaysStatus,
-  getPreMonthDates,
   compareDate,
   getMonthDays,
-  getWeekDate,
-  formatResultDate,
   isEqual,
   getNumTwoBit,
   getWhatDay,
-  getWeekNosOfYear,
 } from '@/utils/date'
+import {
+  getCurrMonthData,
+  getPreMonthDates,
+  getWeekDate,
+  formatResultDate,
+  getDaysStatus,
+  getWeekNosOfYear,
+} from './utils'
 import requestAniFrame from '@/utils/raf'
 import { useConfig } from '@/packages/configprovider'
 import { usePropsValue } from '@/utils/use-props-value'
